Updating Custom Dalli and Dalli
